Move CliArguments and CliParserError to application layer
